TI and Booking Changes

Related work items: #107699

+++ src/main/java/com/dpw/runner/shipment/services/dao/impl/ShipmentDao.java
@@ -302,13 +302,13 @@ public class ShipmentDao implements IShipmentDao {
         // BL# and Reference No can not be repeated
         if(!IsStringNullOrEmpty(request.getHouseBill())) {
             List<ShipmentDetails> shipmentDetails = findByHouseBill(request.getHouseBill());
-            if(shipmentDetails != null && shipmentDetails.size() > 0 && (request.getId() == null || shipmentDetails.get(0).getId() != request.getId())) {
+            if(shipmentDetails != null && shipmentDetails.size() > 0 && (request.getId() == null || shipmentDetails.get(0).getId().longValue() != request.getId().longValue())) {
                 errors.add("Shipment with BL# " + request.getHouseBill() + " already exists.");
             }
         }
         if(!IsStringNullOrEmpty(request.getBookingReference())) {
             List<ShipmentDetails> shipmentDetails = findByBookingReference(request.getBookingReference());
-            if(!shipmentDetails.isEmpty() && (request.getId() == null || shipmentDetails.get(0).getId() != request.getId())) {
+            if(!shipmentDetails.isEmpty() && (request.getId() == null || shipmentDetails.get(0).getId().longValue() != request.getId().longValue())) {
                 errors.add("Shipment with ReferenceNo " + request.getBookingReference() + " already exists.");
             }
         }
